% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_compressor.R
\name{data_compressor}
\alias{data_compressor}
\title{Data Compressor}
\usage{
data_compressor(www_location,
  tar_name = "~/Desktop/eyeIntegration_v100_02.tar.gz", include_DB = T)
}
\arguments{
\item{www_location}{The path which contains the www/ App folder}

\item{tar_name}{path/name_of_tar_file.tar.gz}

\item{include_DB}{Should the SQLite databases be included?}
}
\description{
When given www/ location, tar gzips it for upload to biowulf2
}
\examples{
\dontrun{tar_maker(www_location = paste0(system.file('app', 
package = 'eyeIntegrationApp')), tar_name = 'eyeIntegration_v100_01.tar.gz') }
}
