% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data_for_download.R
\name{extract_data_for_download}
\alias{extract_data_for_download}
\title{Table Maker}
\usage{
extract_data_for_download(app_location,
  EiaD_2017 = "eyeIntegration_human_2017_01.sqlite",
  EiaD_2019 = "EiaD_human_expression_2019_09.sqlite",
  save_dir = "~/Desktop/", version_append)
}
\arguments{
\item{app_location}{The path to the working App (so script can find sqlite 
databases)}

\item{save_dir}{Where to save the tsv files}

\item{version_append}{What to append to file}
}
\description{
Extracts tables for user download to put on biowulf2
}
\examples{
\dontrun{extract_data_for_download(app_location = system.file('app', 
package = 'eyeIntegrationApp'), version_append = 01) }
}
