% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eyeIntegration.R
\name{run_eyeIntegration}
\alias{run_eyeIntegration}
\title{Run eyeIntegration App}
\usage{
run_eyeIntegration()
}
\description{
Wrapper that starts the eyeIntegration Shiny App. You must run 
\code{\link{get_eyeIntegration_datasets}} first.
}
\examples{
\dontrun{run_eyeIntegration()} 
}
