% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eyeIntegration_data.R
\name{get_eyeIntegration_datasets}
\alias{get_eyeIntegration_datasets}
\title{Download eyeIntegration data}
\usage{

  get_eyeIntegration_datasets(file = "eyeIntegration_v102_03_heavyweight.tar.gz",
  destdir = system.file("app", package = "eyeIntegrationApp"),
  destfile = "eyeIntegration_data.tar.gz", method, delete_tar = TRUE,
  verbose = TRUE)
}
\arguments{
\item{file}{The file name on biowulf2}

\item{destdir}{The destination for the eyeIntegration tar.gz file.}

\item{destfile}{The file name for the eyeIntegration tar.gz file}

\item{method}{Download method}

\item{delete_tar}{Delete the tar file after data extraction?}

\item{verbose}{Turn on verbosity for untar()}
}
\description{
Downloads the data underlying eyeIntegration.
}
\examples{
\dontrun{get_eyeIntegration_datasets()} 
}
